Script Creation Overhaul (#20)
* Update restart script creation

- Optimize imports
- Change the script creation in BungeeUpdater.Java and SpigotUpdater.Java to be more similar
- Stop windows Goto loop from being added to the restart script  when the plugin is not running in bungee.  This would cause a second instance of the server to be started.

The loop added for Bungee when a shell script (Linux, MacOS) is being creating still needs to be updated though.

* class and var name changes

* Update README.md

* Remedy script creation for Bungee on linux/ac

Also, the bungee scripts now wait 10 seconds before restarting. Allows the user the chance to stop the server peacefully. Not the ideal solution, but its better for now.

* conflict fix in SpigotUpdater

* Update SpigotUpdater.java

fix braces

* more update

* fix duplicate method call in BungeeUpdater